requirements.txt
# Google Cloud and AI
google-cloud-aiplatform>=1.40.0
google-generativeai>=0.3.2
google-auth>=2.25.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0

# Core Framework
pydantic>=2.5.0
pydantic-settings>=2.1.0
typing-extensions>=4.8.0

# Web API and Communication
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
websockets>=12.0
aiofiles>=23.2.1
httpx>=0.25.0

# Data Processing
numpy>=1.24.0
pandas>=2.1.0
opencv-python>=4.8.1
pillow>=10.1.0

# Environment and Configuration
python-dotenv>=1.0.0
pyyaml>=6.0.1

# Logging and Monitoring
structlog>=23.2.0
rich>=13.7.0

# Message Queuing and Storage
redis>=5.0.0
aioredis>=2.0.1

# Database
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
alembic>=1.13.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Development Tools
black>=23.11.0
ruff>=0.1.6
mypy>=1.7.0

# Additional utilities
tenacity>=8.2.3  # For retry logic
asyncio-throttle>=1.0.2  # For rate limiting
